Ultimos cambios Entregable 2

+++ MrTiendita/Controladores/CrearTicket.cs
@@ -197,7 +197,7 @@ namespace CrearTicketVenta
         }
 
         //Metodo para agregar los totales d ela venta
-        public void AgregarTotales(string texto, decimal total)
+        public void AgregarTotales(string texto, double total)
         {
             //Variables que usaremos
             string resumen, valor, textoCompleto, espacios = "";
@@ -225,7 +225,7 @@ namespace CrearTicketVenta
         }
 
         //Metodo para agreagar articulos al ticket de venta
-        public void AgregaArticulo(string articulo, int cant, decimal precio, decimal importe)
+        public void AgregaArticulo(string articulo, double cant, double precio, double importe)
         {
             //Valida que cant precio e importe esten dentro del rango.
             if (cant.ToString().Length <= 5 && precio.ToString().Length <= 7 && importe.ToString().Length <= 8)
