+++ MrTiendita/Modelos/DTO/Empleado.cs
@@ -12,8 +12,8 @@ namespace MrTiendita.Modelos.DTO
         private String nombre;
         private String a_paterno;
         private String a_materno;
-        private int telefono;
-        private decimal sueldo;
+        private long telefono;
+        private double sueldo;
         private String tipo_empleado;
         private String clave;
 
@@ -21,7 +21,7 @@ namespace MrTiendita.Modelos.DTO
         {
         }
 
-        public Empleado(int id_empleado, string nombre, string a_paterno, string a_materno, int telefono, decimal sueldo, string tipo_empleado, string clave)
+        public Empleado(int id_empleado, string nombre, string a_paterno, string a_materno, long telefono, double sueldo, string tipo_empleado, string clave)
         {
             this.id_empleado = id_empleado;
             this.nombre = nombre;
@@ -37,8 +37,8 @@ namespace MrTiendita.Modelos.DTO
         public string Nombre { get => this.nombre; set => this.nombre = value; }
         public string A_paterno { get => this.a_paterno; set => this.a_paterno = value; }
         public string A_materno { get => this.a_materno; set => this.a_materno = value; }
-        public int Telefono { get => this.telefono; set => this.telefono = value; }
-        public decimal Sueldo { get => this.sueldo; set => this.sueldo = value; }
+        public long Telefono { get => this.telefono; set => this.telefono = value; }
+        public double Sueldo { get => this.sueldo; set => this.sueldo = value; }
         public string Tipo_empleado { get => this.tipo_empleado; set => this.tipo_empleado = value; }
         public string Clave { get => this.clave; set => this.clave = value; }
     }
